+++ src/Plugin.php
@@ -209,7 +209,7 @@ class Plugin extends BasePlugin
     /**
      * @inheritDoc
      */
-    public string $schemaVersion = '4.2.3';
+    public string $schemaVersion = '4.2.4';
 
     /**
      * @inheritdoc