+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
+- Fixed a bug where changing a user’s email would cause extra user elements to be created. ([#3138](https://github.com/craftcms/commerce/issues/3138))
 - Fixed a bug where related sales were showing when creating a new product. 
 - Fixed a bug where Commerce wasn’t invoking `craft\services\Elements::EVENT_AUTHORIZE_*` event handlers.
+- Added `craft\commerce\elements\Order::$orderCompletedEmail`. ([#3138](https://github.com/craftcms/commerce/issues/3138))
 
 ## 4.2.11 - 2023-06-05